+++ clang/include/clang/Parse/Parser.h
@@ -7168,13 +7168,16 @@ public:
     AllowStandaloneOpenMPDirectives = 0x2,
     /// This context is at the top level of a GNU statement expression.
     InStmtExpr = 0x4,
+    /// This context is the C99 secondary-block in selection or iteration
+    /// statement.
+    SecondaryBlockInC = 0x8,
 
     /// The context of a regular substatement.
     SubStmt = 0,
     /// The context of a compound-statement.
     Compound = AllowDeclarationsInC | AllowStandaloneOpenMPDirectives,
 
-    LLVM_MARK_AS_BITMASK_ENUM(InStmtExpr)
+    LLVM_MARK_AS_BITMASK_ENUM(SecondaryBlockInC)
   };
 
   /// Act on an expression statement that might be the last statement in a